CTFNotifier_Discord_Bot_v2/requirements.txt
discord.py==2.3.1
python-dotenv==1.0.0
pytz==2022.1
Requests==2.31.0
aiosqlite==0.19.0

